@@ -18,7 +18,7 @@
   "homepage": "https://github.com/cityssm/lottery-licence-manager#readme",
   "dependencies": {
     "@cityssm/bulma-webapp-js": "^0.4.2",
-    "@cityssm/expressjs-server-js": "^0.6.0",
+    "@cityssm/expressjs-server-js": "^0.6.1",
     "@fortawesome/fontawesome-free": "^5.14.0",
     "bcrypt": "^5.0.0",
     "better-sqlite3": "^7.1.0",
@@ -51,8 +51,8 @@
     "@types/marked": "^1.1.0",
     "@types/mocha": "^8.0.3",
     "@types/morgan": "^1.9.1",
-    "@typescript-eslint/eslint-plugin": "^3.9.1",
-    "@typescript-eslint/parser": "^3.9.1",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
     "eslint": "^7.7.0",
     "eslint-config-standard-with-typescript": "^18.0.2",
     "eslint-plugin-import": "^2.22.0",
@@ -62,8 +62,8 @@
     "github-markdown-css": "^4.0.0",
     "gulp": "^4.0.2",
     "gulp-minify": "^3.1.0",
-    "mocha": "^8.1.1",
+    "mocha": "^8.1.2",
     "nyc": "^15.1.0",
-    "typescript": "^3.9.7"
+    "typescript": "^4.0.2"
   }
 }
