@@ -1,17 +1,25 @@
 import { Router } from "express";
 
-import * as path from "path";
-import * as ejs from "ejs";
-
 import * as dateTimeFns from "@cityssm/expressjs-server-js/dateTimeFns";
 import * as configFns from "../helpers/configFns";
 
-import * as licencesDB from "../helpers/licencesDB";
-import * as licencesDBOrganizations from "../helpers/licencesDB-organizations";
-import { Organization } from "../helpers/llmTypes";
+import * as permissionHandlers from "../handlers/permissions";
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const convertHTMLToPDF = require("pdf-puppeteer");
+import * as handler_doSearch from "../handlers/licences-post/doSearch";
+
+import * as handler_view from "../handlers/licences-get/view";
+import * as handler_new from "../handlers/licences-get/new";
+import * as handler_edit from "../handlers/licences-get/edit";
+import * as handler_print from "../handlers/licences-get/print";
+
+import * as handler_doSave from "../handlers/licences-post/doSave";
+import * as handler_doUnissueLicence from "../handlers/licences-post/doUnissueLicence";
+
+import * as handler_doAddTransaction from "../handlers/licences-post/doAddTransaction";
+import * as handler_doVoidTransaction from "../handlers/licences-post/doVoidTransaction";
+
+import * as licencesDB from "../helpers/licencesDB";
+
 
 const router = Router();
 
@@ -30,15 +38,7 @@ router.get("/", (_req, res) => {
 });
 
 
-router.post("/doSearch", (req, res) => {
-
-  res.json(licencesDB.getLicences(req.body, req.session, {
-    includeOrganization: true,
-    limit: req.body.limit,
-    offset: req.body.offset
-  }));
-
-});
+router.post("/doSearch", handler_doSearch.handler);
 
 
 /*
@@ -134,71 +134,9 @@ router.post("/doGetActiveLicenceSummary", (req, res) => {
 router.get([
   "/new",
   "/new/:organizationID"
-], (req, res) => {
-
-  // Check permission
-
-  if (!req.session.user.userProperties.canCreate) {
-
-    res.redirect("/licences/?error=accessDenied");
-    return;
-
-  }
-
-  // Get organization (if set)
-
-  const organizationID = parseInt(req.params.organizationID, 10);
-
-  let organization: Organization = null;
-
-  if (organizationID) {
-
-    organization = licencesDBOrganizations.getOrganization(organizationID, req.session);
-
-    if (organization && !organization.isEligibleForLicences) {
-
-      organization = null;
-
-    }
-
-  }
-
-  // Use current date as default
-
-  const currentDateAsString = dateTimeFns.dateToString(new Date());
-
-  // Get next external licence number
-
-  let externalLicenceNumber = "";
-
-  const licenceNumberCalculationType = configFns.getProperty("licences.externalLicenceNumber.newCalculation");
-
-  if (licenceNumberCalculationType === "range") {
-
-    externalLicenceNumber = licencesDB.getNextExternalLicenceNumberFromRange().toString();
-
-  }
-
-  res.render("licence-edit", {
-    headTitle: "Licence Create",
-    isCreate: true,
-    licence: {
-      externalLicenceNumber,
-      applicationDateString: currentDateAsString,
-      municipality: configFns.getProperty("defaults.city"),
-      startDateString: currentDateAsString,
-      endDateString: currentDateAsString,
-      startTimeString: "00:00",
-      endTimeString: "00:00",
-      licenceDetails: "",
-      termsConditions: "",
-      licenceTicketTypes: [],
-      events: []
-    },
-    organization
-  });
-
-});
+],
+  permissionHandlers.createGetHandler,
+  handler_new.handler);
 
 
 router.post("/doGetDistinctTermsConditions", (req, res) => {
@@ -229,114 +167,19 @@ router.post("/doGetTicketTypes", (req, res) => {
 });
 
 
-router.post("/doSave", (req, res) => {
-
-  if (!req.session.user.userProperties.canCreate) {
-
-    res
-      .status(403)
-      .json({
-        success: false,
-        message: "Forbidden"
-      });
-
-    return;
-
-  }
-
-  if (req.body.licenceID === "") {
-
-    const newLicenceID = licencesDB.createLicence(req.body, req.session);
-
-    res.json({
-      success: true,
-      licenceID: newLicenceID
-    });
-
-  } else {
-
-    const changeCount = licencesDB.updateLicence(req.body, req.session);
-
-    if (changeCount) {
-
-      res.json({
-        success: true,
-        message: "Licence updated successfully."
-      });
-
-    } else {
-
-      res.json({
-        success: false,
-        message: "Record Not Saved"
-      });
-
-    }
-
-  }
-
-});
+router.post("/doSave",
+  permissionHandlers.createPostHandler,
+  handler_doSave.handler);
 
 
-router.post("/doAddTransaction", (req, res) => {
-
-  if (!req.session.user.userProperties.canCreate) {
-
-    res
-      .status(403)
-      .json({
-        success: false,
-        message: "Forbidden"
-      });
-    return;
-
-  }
-
-  const newTransactionIndex = licencesDB.addTransaction(req.body, req.session);
-
-  res.json({
-    success: true,
-    message: "Transaction Added Successfully",
-    transactionIndex: newTransactionIndex
-  });
-
-});
+router.post("/doAddTransaction",
+  permissionHandlers.createPostHandler,
+  handler_doAddTransaction.handler);
 
 
-router.post("/doVoidTransaction", (req, res) => {
-
-  if (!req.session.user.userProperties.canCreate) {
-
-    res
-      .status(403)
-      .json({
-        success: false,
-        message: "Forbidden"
-      });
-
-    return;
-
-  }
-
-  const success = licencesDB.voidTransaction(req.body.licenceID, req.body.transactionIndex, req.session);
-
-  if (success) {
-
-    res.json({
-      success: true,
-      message: "Transaction Voided Successfully"
-    });
-
-  } else {
-
-    res.json({
-      success: false,
-      message: "Transaction Not Voided"
-    });
-
-  }
-
-});
+router.post("/doVoidTransaction",
+  permissionHandlers.createPostHandler,
+  handler_doVoidTransaction.handler);
 
 
 router.post("/doIssueLicence", (req, res) => {
@@ -375,39 +218,9 @@ router.post("/doIssueLicence", (req, res) => {
 });
 
 
-router.post("/doUnissueLicence", (req, res) => {
-
-  if (!req.session.user.userProperties.canCreate) {
-
-    res
-      .status(403)
-      .json({
-        success: false,
-        message: "Forbidden"
-      });
-    return;
-
-  }
-
-  const success = licencesDB.unissueLicence(req.body.licenceID, req.session);
-
-  if (success) {
-
-    res.json({
-      success: true,
-      message: "Licence Unissued Successfully"
-    });
-
-  } else {
-
-    res.json({
-      success: false,
-      message: "Licence Not Unissued"
-    });
-
-  }
-
-});
+router.post("/doUnissueLicence",
+  permissionHandlers.createPostHandler,
+  handler_doUnissueLicence.handler);
 
 
 router.post("/doDelete", (req, res) => {
@@ -457,131 +270,13 @@ router.post("/doDelete", (req, res) => {
 });
 
 
-router.get("/:licenceID", (req, res) => {
-
-  const licenceID = parseInt(req.params.licenceID, 10);
-
-  const licence = licencesDB.getLicence(licenceID, req.session);
-
-  if (!licence) {
-
-    res.redirect("/licences/?error=licenceNotFound");
-    return;
-
-  }
-
-  const organization = licencesDBOrganizations.getOrganization(licence.organizationID, req.session);
-
-  const headTitle =
-    configFns.getProperty("licences.externalLicenceNumber.isPreferredID")
-      ? "Licence " + licence.externalLicenceNumber
-      : "Licence #" + licenceID.toString();
-
-  res.render("licence-view", {
-    headTitle,
-    licence,
-    organization
-  });
-
-});
+router.get("/:licenceID", handler_view.handler);
 
 
-router.get("/:licenceID/edit", (req, res) => {
-
-  const licenceID = parseInt(req.params.licenceID, 10);
-
-  if (!req.session.user.userProperties.canCreate) {
-
-    res.redirect("/licences/" + licenceID.toString() + "/?error=accessDenied");
-    return;
-
-  }
-
-  const licence = licencesDB.getLicence(licenceID, req.session);
-
-  if (!licence) {
-
-    res.redirect("/licences/?error=licenceNotFound");
-    return;
-
-  } else if (!licence.canUpdate) {
-
-    res.redirect("/licences/" + licenceID.toString() + "/?error=accessDenied");
-    return;
-
-  }
+router.get("/:licenceID/edit", handler_edit.handler);
 
 
-  const organization = licencesDBOrganizations.getOrganization(licence.organizationID, req.session);
-
-  const feeCalculation = configFns.getProperty("licences.feeCalculationFn")(licence);
-
-  res.render("licence-edit", {
-    headTitle: "Licence #" + licenceID.toString() + " Update",
-    isCreate: false,
-    licence,
-    organization,
-    feeCalculation
-  });
-
-});
-
-
-router.get("/:licenceID/print", (req, res, next) => {
-
-  const licenceID = parseInt(req.params.licenceID, 10);
-
-  const licence = licencesDB.getLicence(licenceID, req.session);
-
-  if (!licence) {
-
-    res.redirect("/licences/?error=licenceNotFound");
-    return;
-
-  }
-
-  if (!licence.issueDate) {
-
-    res.redirect("/licences/?error=licenceNotIssued");
-    return;
-
-  }
-
-  const organization = licencesDBOrganizations.getOrganization(licence.organizationID, req.session);
-
-  ejs.renderFile(
-    path.join(__dirname, "../reports/", configFns.getProperty("licences.printTemplate")), {
-      configFns,
-      licence,
-      organization
-    }, {},
-    (ejsErr, ejsData) => {
-
-      if (ejsErr) {
-        return next(ejsErr);
-      }
-
-      convertHTMLToPDF(ejsData, (pdf) => {
-
-        res.setHeader("Content-Disposition",
-          "attachment;" +
-          " filename=licence-" + licenceID.toString() + "-" + licence.recordUpdate_timeMillis.toString() + ".pdf"
-        );
-
-        res.setHeader("Content-Type", "application/pdf");
-
-        res.send(pdf);
-
-      }, {
-          format: "Letter",
-          printBackground: true,
-          preferCSSPageSize: true
-        });
-
-      return null;
-    }
-  );
-});
+router.get("/:licenceID/print", handler_print.handler);
 
 
 router.get("/:licenceID/poke", (req, res) => {
