@@ -1,117 +1,11 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getInactiveLocations = exports.mergeLocations = exports.restoreLocation = exports.deleteLocation = exports.updateLocation = exports.createLocation = exports.getLocation = exports.getLocations = void 0;
-const licencesDB_1 = require("./licencesDB");
+exports.getInactiveLocations = exports.mergeLocations = exports.restoreLocation = exports.deleteLocation = exports.updateLocation = exports.createLocation = void 0;
 const sqlite = require("better-sqlite3");
+const databasePaths_1 = require("../data/databasePaths");
 const dateTimeFns = require("@cityssm/expressjs-server-js/dateTimeFns");
-exports.getLocations = (reqSession, queryOptions) => {
-    const db = sqlite(licencesDB_1.dbPath, {
-        readonly: true
-    });
-    const sqlParams = [];
-    let sqlWhereClause = " where lo.recordDelete_timeMillis is null";
-    if (queryOptions.locationNameAddress && queryOptions.locationNameAddress !== "") {
-        const locationNameAddressSplit = queryOptions.locationNameAddress.toLowerCase().split(" ");
-        for (const locationPiece of locationNameAddressSplit) {
-            sqlWhereClause += " and (instr(lower(lo.locationName), ?) or instr(lower(lo.locationAddress1),?))";
-            sqlParams.push(locationPiece);
-            sqlParams.push(locationPiece);
-        }
-    }
-    if (queryOptions.locationIsDistributor !== -1) {
-        sqlWhereClause += " and lo.locationIsDistributor = ?";
-        sqlParams.push(queryOptions.locationIsDistributor);
-    }
-    if (queryOptions.locationIsManufacturer !== -1) {
-        sqlWhereClause += " and lo.locationIsManufacturer = ?";
-        sqlParams.push(queryOptions.locationIsManufacturer);
-    }
-    if (queryOptions.locationIsActive !== "") {
-        const currentDate = dateTimeFns.dateToInteger(new Date());
-        sqlWhereClause += " and (lo.locationID in (" +
-            "select lx.locationID from LotteryLicences lx" +
-            " where lx.recordDelete_timeMillis is null" +
-            " and lx.issueDate is not null and lx.endDate >= ?))";
-        sqlParams.push(currentDate);
-    }
-    let count = 0;
-    if (queryOptions.limit !== -1) {
-        count = db.prepare("select ifnull(count(*), 0) as cnt" +
-            " from Locations lo" +
-            sqlWhereClause)
-            .get(sqlParams)
-            .cnt;
-    }
-    let sql = "select lo.locationID, lo.locationName," +
-        " lo.locationAddress1, lo.locationAddress2, lo.locationCity, lo.locationProvince," +
-        " lo.locationIsDistributor, lo.locationIsManufacturer," +
-        " l.licences_endDateMax, coalesce(l.licences_count, 0) as licences_count," +
-        " d.distributor_endDateMax, coalesce(d.distributor_count, 0) as distributor_count," +
-        " m.manufacturer_endDateMax, coalesce(m.manufacturer_count, 0) as manufacturer_count" +
-        " from Locations lo" +
-        (" left join (" +
-            "select locationID," +
-            " count(licenceID) as licences_count, max(endDate) as licences_endDateMax" +
-            " from LotteryLicences" +
-            " where recordDelete_timeMillis is null" +
-            " group by locationID" +
-            ") l on lo.locationID = l.locationID") +
-        (" left join (" +
-            "select t.distributorLocationID," +
-            " count(*) as distributor_count, max(l.endDate) as distributor_endDateMax" +
-            " from LotteryLicenceTicketTypes t" +
-            " left join LotteryLicences l on t.licenceID = l.licenceID" +
-            " where t.recordDelete_timeMillis is null" +
-            " group by t.distributorLocationID" +
-            ") d on lo.locationID = d.distributorLocationID") +
-        (" left join (" +
-            "select t.manufacturerLocationID," +
-            " count(*) as manufacturer_count, max(l.endDate) as manufacturer_endDateMax" +
-            " from LotteryLicenceTicketTypes t" +
-            " left join LotteryLicences l on t.licenceID = l.licenceID" +
-            " where t.recordDelete_timeMillis is null" +
-            " group by t.manufacturerLocationID" +
-            ") m on lo.locationID = m.manufacturerLocationID") +
-        sqlWhereClause +
-        " order by case when lo.locationName = '' then lo.locationAddress1 else lo.locationName end";
-    if (queryOptions.limit !== -1) {
-        sql += " limit " + queryOptions.limit.toString() +
-            " offset " + queryOptions.offset.toString();
-    }
-    const rows = db.prepare(sql).all(sqlParams);
-    db.close();
-    for (const ele of rows) {
-        ele.recordType = "location";
-        ele.locationDisplayName =
-            ele.locationName === "" ? ele.locationAddress1 : ele.locationName;
-        ele.licences_endDateMaxString = dateTimeFns.dateIntegerToString(ele.licences_endDateMax);
-        ele.distributor_endDateMaxString = dateTimeFns.dateIntegerToString(ele.distributor_endDateMax);
-        ele.manufacturer_endDateMaxString = dateTimeFns.dateIntegerToString(ele.manufacturer_endDateMax);
-        ele.canUpdate = licencesDB_1.canUpdateObject(ele, reqSession);
-    }
-    return {
-        count: (queryOptions.limit === -1 ? rows.length : count),
-        locations: rows
-    };
-};
-exports.getLocation = (locationID, reqSession) => {
-    const db = sqlite(licencesDB_1.dbPath, {
-        readonly: true
-    });
-    const locationObj = db.prepare("select * from Locations" +
-        " where locationID = ?")
-        .get(locationID);
-    if (locationObj) {
-        locationObj.recordType = "location";
-        locationObj.canUpdate = licencesDB_1.canUpdateObject(locationObj, reqSession);
-    }
-    db.close();
-    locationObj.locationDisplayName =
-        locationObj.locationName === "" ? locationObj.locationAddress1 : locationObj.locationName;
-    return locationObj;
-};
 exports.createLocation = (reqBody, reqSession) => {
-    const db = sqlite(licencesDB_1.dbPath);
+    const db = sqlite(databasePaths_1.licencesDB);
     const nowMillis = Date.now();
     const info = db.prepare("insert into Locations" +
         " (locationName, locationAddress1, locationAddress2, locationCity, locationProvince, locationPostalCode," +
@@ -123,7 +17,7 @@ exports.createLocation = (reqBody, reqSession) => {
     return Number(info.lastInsertRowid);
 };
 exports.updateLocation = (reqBody, reqSession) => {
-    const db = sqlite(licencesDB_1.dbPath);
+    const db = sqlite(databasePaths_1.licencesDB);
     const nowMillis = Date.now();
     const info = db.prepare("update Locations" +
         " set locationName = ?," +
@@ -143,7 +37,7 @@ exports.updateLocation = (reqBody, reqSession) => {
     return info.changes > 0;
 };
 exports.deleteLocation = (locationID, reqSession) => {
-    const db = sqlite(licencesDB_1.dbPath);
+    const db = sqlite(databasePaths_1.licencesDB);
     const nowMillis = Date.now();
     const info = db.prepare("update Locations" +
         " set recordDelete_userName = ?," +
@@ -155,7 +49,7 @@ exports.deleteLocation = (locationID, reqSession) => {
     return info.changes > 0;
 };
 exports.restoreLocation = (locationID, reqSession) => {
-    const db = sqlite(licencesDB_1.dbPath);
+    const db = sqlite(databasePaths_1.licencesDB);
     const nowMillis = Date.now();
     const info = db.prepare("update Locations" +
         " set recordDelete_userName = null," +
@@ -169,7 +63,7 @@ exports.restoreLocation = (locationID, reqSession) => {
     return info.changes > 0;
 };
 exports.mergeLocations = (targetLocationID, sourceLocationID, reqSession) => {
-    const db = sqlite(licencesDB_1.dbPath);
+    const db = sqlite(databasePaths_1.licencesDB);
     const nowMillis = Date.now();
     const locationAttributes = db.prepare("select max(locationIsDistributor) as locationIsDistributorMax," +
         " max(locationIsManufacturer) as locationIsManufacturerMax," +
@@ -218,7 +112,7 @@ exports.getInactiveLocations = (inactiveYears) => {
     const cutoffDate = new Date();
     cutoffDate.setFullYear(cutoffDate.getFullYear() - inactiveYears);
     const cutoffDateInteger = dateTimeFns.dateToInteger(cutoffDate);
-    const db = sqlite(licencesDB_1.dbPath, {
+    const db = sqlite(databasePaths_1.licencesDB, {
         readonly: true
     });
     const rows = db.prepare("select lo.locationID, lo.locationName, lo.locationAddress1," +
