@@ -1,17 +1,7 @@
 /// <reference types="express-serve-static-core" />
 /// <reference types="compression" />
 /// <reference types="express-session" />
-import * as llm from "./llmTypes";
-export declare const getOrganizations: (reqBody: {
-    organizationName?: string;
-    representativeName?: string;
-    isEligibleForLicences?: string;
-    organizationIsActive?: string;
-}, reqSession: Express.SessionData, includeOptions: {
-    limit: number;
-    offset?: number;
-}) => llm.Organization[];
-export declare const getOrganization: (organizationID: number, reqSession: Express.SessionData) => llm.Organization;
+import type * as llm from "../types/recordTypes";
 export declare const createOrganization: (reqBody: llm.Organization, reqSession: Express.SessionData) => number;
 export declare const updateOrganization: (reqBody: llm.Organization, reqSession: Express.SessionData) => boolean;
 export declare const deleteOrganization: (organizationID: number, reqSession: Express.SessionData) => boolean;
@@ -22,13 +12,10 @@ export declare const addOrganizationRepresentative: (organizationID: number, req
 export declare const updateOrganizationRepresentative: (organizationID: number, reqBody: llm.OrganizationRepresentative) => llm.OrganizationRepresentative;
 export declare const deleteOrganizationRepresentative: (organizationID: number, representativeIndex: number) => boolean;
 export declare const setDefaultOrganizationRepresentative: (organizationID: number, representativeIndex: number) => boolean;
-export declare const getOrganizationRemarks: (organizationID: number, reqSession: Express.SessionData) => llm.OrganizationRemark[];
 export declare const getOrganizationRemark: (organizationID: number, remarkIndex: number, reqSession: Express.SessionData) => llm.OrganizationRemark;
 export declare const addOrganizationRemark: (reqBody: llm.OrganizationRemark, reqSession: Express.SessionData) => number;
 export declare const updateOrganizationRemark: (reqBody: llm.OrganizationRemark, reqSession: Express.SessionData) => boolean;
 export declare const deleteOrganizationRemark: (organizationID: number, remarkIndex: number, reqSession: Express.SessionData) => boolean;
-export declare const getUndismissedOrganizationReminders: (reqSession: Express.SessionData) => llm.OrganizationReminder[];
-export declare const getOrganizationReminders: (organizationID: number, reqSession: Express.SessionData) => llm.OrganizationReminder[];
 export declare const getOrganizationReminder: (organizationID: number, reminderIndex: number, reqSession: Express.SessionData) => llm.OrganizationReminder;
 export declare const addOrganizationReminder: (reqBody: {
     organizationID: string;
