@@ -1,153 +1,14 @@
-import { dbPath, canUpdateObject } from "./licencesDB";
+import { canUpdateObject } from "./licencesDB";
 import * as sqlite from "better-sqlite3";
 
-import * as llm from "./llmTypes";
+import { licencesDB as dbPath } from "../data/databasePaths";
+
 import * as dateTimeFns from "@cityssm/expressjs-server-js/dateTimeFns";
+import type * as llm from "../types/recordTypes";
 
 /*
  * ORGANIZATIONS
  */
-
-export const getOrganizations = (reqBody: {
-  organizationName?: string;
-  representativeName?: string;
-  isEligibleForLicences?: string;
-  organizationIsActive?: string;
-}, reqSession: Express.SessionData, includeOptions: {
-  limit: number;
-  offset?: number;
-}) => {
-
-  const db = sqlite(dbPath, {
-    readonly: true
-  });
-
-  const sqlParams: Array<string | number> = [dateTimeFns.dateToInteger(new Date())];
-
-  let sql = "select o.organizationID, o.organizationName, o.isEligibleForLicences, o.organizationNote," +
-    " r.representativeName," +
-    " sum(case when l.endDate >= ? then 1 else 0 end) as licences_activeCount," +
-    " max(l.endDate) as licences_endDateMax," +
-    " o.recordCreate_userName, o.recordCreate_timeMillis, o.recordUpdate_userName, o.recordUpdate_timeMillis" +
-    " from Organizations o" +
-    " left join OrganizationRepresentatives r on o.organizationID = r.organizationID and r.isDefault = 1" +
-    " left join LotteryLicences l on o.organizationID = l.organizationID and l.recordDelete_timeMillis is null" +
-    " where o.recordDelete_timeMillis is null";
-
-  if (reqBody.organizationName && reqBody.organizationName !== "") {
-
-    const organizationNamePieces = reqBody.organizationName.toLowerCase().split(" ");
-
-    for (const organizationPiece of organizationNamePieces) {
-
-      sql += " and instr(lower(o.organizationName), ?)";
-      sqlParams.push(organizationPiece);
-    }
-  }
-
-  if (reqBody.representativeName && reqBody.representativeName !== "") {
-
-    const representativeNamePieces = reqBody.representativeName.toLowerCase().split(" ");
-
-    for (const representativePiece of representativeNamePieces) {
-
-      sql += " and o.organizationID in (" +
-        "select organizationID from OrganizationRepresentatives where instr(lower(representativeName), ?)" +
-        ")";
-
-      sqlParams.push(representativePiece);
-    }
-  }
-
-  if (reqBody.isEligibleForLicences && reqBody.isEligibleForLicences !== "") {
-
-    sql += " and o.isEligibleForLicences = ?";
-    sqlParams.push(reqBody.isEligibleForLicences);
-
-  }
-
-  if (reqBody.organizationIsActive && reqBody.organizationIsActive !== "") {
-
-    const currentDate = dateTimeFns.dateToInteger(new Date());
-
-    sql += " and o.organizationID in (" +
-      "select lx.organizationID from LotteryLicences lx" +
-      " where lx.recordDelete_timeMillis is null" +
-      " and lx.issueDate is not null and lx.endDate >= ?)";
-
-    sqlParams.push(currentDate);
-  }
-
-  sql += " group by o.organizationID, o.organizationName, o.isEligibleForLicences, o.organizationNote," +
-    " r.representativeName," +
-    " o.recordCreate_userName, o.recordCreate_timeMillis, o.recordUpdate_userName, o.recordUpdate_timeMillis" +
-    " order by o.organizationName, o.organizationID";
-
-  if (includeOptions.limit !== -1) {
-
-    sql += " limit " + includeOptions.limit.toString() +
-      " offset " + includeOptions.offset.toString();
-
-  }
-
-  const rows: llm.Organization[] = db.prepare(sql).all(sqlParams);
-
-  db.close();
-
-  for (const ele of rows) {
-
-    ele.recordType = "organization";
-
-    ele.licences_endDateMaxString = dateTimeFns.dateIntegerToString(ele.licences_endDateMax || 0);
-
-    ele.canUpdate = canUpdateObject(ele, reqSession);
-
-    delete ele.recordCreate_userName;
-    delete ele.recordCreate_timeMillis;
-    delete ele.recordUpdate_userName;
-    delete ele.recordUpdate_timeMillis;
-  }
-
-  return rows;
-
-};
-
-
-export const getOrganization = (organizationID: number, reqSession: Express.SessionData): llm.Organization => {
-
-  const db = sqlite(dbPath, {
-    readonly: true
-  });
-
-  const organizationObj: llm.Organization =
-    db.prepare("select * from Organizations" +
-      " where organizationID = ?")
-      .get(organizationID);
-
-  if (organizationObj) {
-
-    organizationObj.recordType = "organization";
-
-    organizationObj.fiscalStartDateString = dateTimeFns.dateIntegerToString(organizationObj.fiscalStartDate);
-    organizationObj.fiscalEndDateString = dateTimeFns.dateIntegerToString(organizationObj.fiscalEndDate);
-
-    organizationObj.canUpdate = canUpdateObject(organizationObj, reqSession);
-
-    const representativesList: llm.OrganizationRepresentative[] =
-      db.prepare("select * from OrganizationRepresentatives" +
-        " where organizationID = ?" +
-        " order by isDefault desc, representativeName")
-        .all(organizationID);
-
-    organizationObj.organizationRepresentatives = representativesList;
-
-  }
-
-  db.close();
-
-  return organizationObj;
-
-};
 
 
 /**
@@ -502,39 +363,6 @@ export const setDefaultOrganizationRepresentative = (organizationID: number, rep
  */
 
 
-export const getOrganizationRemarks = (organizationID: number, reqSession: Express.SessionData) => {
-
-  const db = sqlite(dbPath, {
-    readonly: true
-  });
-
-  const remarks: llm.OrganizationRemark[] =
-    db.prepare("select remarkIndex," +
-      " remarkDate, remarkTime," +
-      " remark, isImportant," +
-      " recordCreate_userName, recordCreate_timeMillis, recordUpdate_userName, recordUpdate_timeMillis" +
-      " from OrganizationRemarks" +
-      " where recordDelete_timeMillis is null" +
-      " and organizationID = ?" +
-      " order by remarkDate desc, remarkTime desc")
-      .all(organizationID);
-
-  db.close();
-
-  for (const remark of remarks) {
-
-    remark.recordType = "remark";
-
-    remark.remarkDateString = dateTimeFns.dateIntegerToString(remark.remarkDate || 0);
-    remark.remarkTimeString = dateTimeFns.timeIntegerToString(remark.remarkTime || 0);
-
-    remark.canUpdate = canUpdateObject(remark, reqSession);
-  }
-
-  return remarks;
-};
-
-
 export const getOrganizationRemark =
   (organizationID: number, remarkIndex: number, reqSession: Express.SessionData) => {
 
@@ -674,73 +502,6 @@ export const deleteOrganizationRemark =
 /*
  * Organization Reminders
  */
-
-
-export const getUndismissedOrganizationReminders = (reqSession: Express.SessionData) => {
-
-  const db = sqlite(dbPath, {
-    readonly: true
-  });
-
-  const reminders: llm.OrganizationReminder[] =
-    db.prepare("select r.organizationID, o.organizationName, r.reminderIndex," +
-      " r.reminderTypeKey, r.reminderDate," +
-      " r.reminderStatus, r.reminderNote," +
-      " r.recordUpdate_userName, r.recordUpdate_timeMillis" +
-      " from OrganizationReminders r" +
-      " left join Organizations o on r.organizationID = o.organizationID" +
-      " where r.recordDelete_timeMillis is null" +
-      " and o.recordDelete_timeMillis is null" +
-      " and r.dismissedDate is null" +
-      " order by r.reminderDate, o.organizationName, r.reminderTypeKey")
-      .all();
-
-  db.close();
-
-  for (const reminder of reminders) {
-
-    reminder.recordType = "reminder";
-
-    reminder.reminderDateString = dateTimeFns.dateIntegerToString(reminder.reminderDate || 0);
-
-    reminder.canUpdate = canUpdateObject(reminder, reqSession);
-  }
-
-  return reminders;
-};
-
-
-export const getOrganizationReminders = (organizationID: number, reqSession: Express.SessionData) => {
-
-  const db = sqlite(dbPath, {
-    readonly: true
-  });
-
-  const reminders: llm.OrganizationReminder[] =
-    db.prepare("select reminderIndex," +
-      " reminderTypeKey, reminderDate, dismissedDate," +
-      " reminderStatus, reminderNote," +
-      " recordCreate_userName, recordCreate_timeMillis, recordUpdate_userName, recordUpdate_timeMillis" +
-      " from OrganizationReminders" +
-      " where recordDelete_timeMillis is null" +
-      " and organizationID = ?" +
-      " order by case when dismissedDate is null then 0 else 1 end, reminderDate desc, dismissedDate desc")
-      .all(organizationID);
-
-  db.close();
-
-  for (const reminder of reminders) {
-
-    reminder.recordType = "reminder";
-
-    reminder.reminderDateString = dateTimeFns.dateIntegerToString(reminder.reminderDate || 0);
-    reminder.dismissedDateString = dateTimeFns.dateIntegerToString(reminder.dismissedDate || 0);
-
-    reminder.canUpdate = canUpdateObject(reminder, reqSession);
-  }
-
-  return reminders;
-};
 
 
 export const getOrganizationReminder =
