@@ -1,5 +1,5 @@
 import type { cityssmGlobal } from "@cityssm/bulma-webapp-js/src/types";
-import type * as llmTypes from "../../helpers/llmTypes";
+import type * as llmTypes from "../../types/recordTypes";
 
 declare const cityssm: cityssmGlobal;
 
@@ -131,7 +131,7 @@ declare const cityssm: cityssmGlobal;
           (responseJSON: { success: boolean }) => {
 
             if (responseJSON.success) {
-              window.location.reload(true);
+              window.location.reload();
 
             } else {
               cityssm.alertModal("Merge Not Completed", "Please try again.", "OK", "danger");
